+++ lib/src/inverted_index/async_index.dart
@@ -73,7 +73,7 @@ class AsyncCallbackIndex
       required this.postingsLoader,
       required this.postingsUpdater,
       required this.tokenizer,
-      this.k = 3,
+      this.k = 2,
       this.zones = const <String, double>{},
       this.phraseLength = 1})
       : assert(phraseLength > 0, 'The phrase length must be 1 or greater');